common: cli_hush: terminate all commands when Ctrl-C is pressed
Check Ctrl-C status before executing each command in the pipeline so
that the Ctrl-C can skip all subsequent commands. in the meantime, we
need to clear the ctrlc status at the end of a pipeline. This should
be the default behavior of most shell parsers.

Before the change:
```
=> echo start; while true; do echo loop; sleep 1; done; echo int; if true; then echo if; fi; echo stop
start
loop
loop	<-- Input Ctrl-C
int
if
stop
=>
```
After the change:
```
=> echo start; while true; do echo loop; sleep 1; done; echo int; if true; then echo if; fi; echo stop
start
loop
loop	<-- Input Ctrl-C
^C
=>
```

Signed-off-by: Shiji Yang <yangshiji66@outlook.com>